backend/requirements.txt
fastapi>=0.134.0,<0.136
starlette>=0.49.1,<0.50
uvicorn[standard]>=0.32.0,<0.33
python-dotenv>=1.0.0,<2.0
requests>=2.32.0,<3.0
httpx>=0.27.0,<0.29
pydantic>=2.0,<3.0
slowapi>=0.1.9
python-json-logger>=2.0.7
